Feature/auth logic/nathan (#26)
* feat: Add SignUpUseCase, User model, and update User and Wallet entities

- Add SignUpUseCase with execute function
- Add User model
- Update UserEntity: make name and surname nullable
- Update WalletEntity: add currency field

* Implement SignUp feature

- Added `SignUpUseCase` to handle user sign-up logic.
- Created `User` data class to represent user data.
- Updated `SignupViewModel` to interact with `SignUpUseCase`.
- Updated `SignupFragment` to collect user input and trigger sign-up via `SignupViewModel`.

* Refactor: Rename Signup to signUp in SignupFragment

Renamed the `Signup` function to `signUp` in the `SignupFragment`'s ViewModel interaction.

* feat: Implement Hilt for dependency injection and setup signup feature

- Integrated Hilt for dependency injection across the application, including Activities, Fragments, and ViewModels.
- Added necessary Hilt dependencies and plugins to the project.
- Created `AppModule` to provide dependencies like `SignUpUseCase`, `UserDao`, and `AppDatabase`
- Implemented `SignUpUseCase` to interact with the database for user registration.
- Modified `SignupViewModel` to utilize `SignUpUseCase` and launch a coroutine for handling the registration process.
- Updated `SignupFragment` and `AuthActivity` to be Hilt entry points.
- Created `AppDatabase` and `UserDao` and `UserEntity` to store the users on a local database.
- Updated the transaction Id and the user id to be a String.

* Refactor: Relocate entities and DAOs to correct packages

This commit moves the entity classes and their corresponding DAOs to the correct packages. It also updates the imports and references in the affected classes.
- `SignupFragment` and `SignupViewModel` moved to `data/local` and `data/entity` respectively.
- Entity and Dao classes were moved from `data/entity` and `data/local` to `data/entity` and `data/local`.
- Imports updated in the `AuthActivity` and `SignUpUseCase`.
- Removed `SignUpUseCase` from `AppModule`.

* Refactor: Rename userId to user_id in SignupViewModel

Rename `userId` to `user_id` within the `UserEntity` instantiation in `SignupViewModel.kt`.
Add `println` to see the inserted data into the database in `SignUpUseCase.kt`

* Refactor: Move SignupFragment and SignupViewModel to presentation layer

This commit moves `SignupFragment` and `SignupViewModel` from the `data.local.signup` and `data.entity` packages respectively to the `presentation.ui.auth.signup` package. It also updates the package names to reflect this change. Additionally, updates `SignupUseCase` with a trailing line. Finally, it updates `fragment_signup.xml` to reflect the package refactor.

* Implement user sign-up functionality

- Added the ability to create a UserEntity object in `SignupViewModel` and call the `SignUpUseCase`
- Implemented `SignUpUseCase` to insert the user into the database
- Updated `SignupFragment` to collect user input and pass it to `SignupViewModel`
- Added println for debugging purposes.

* Implement user input validation in SignupFragment

The commit adds user input validation to the SignupFragment to ensure that the email, password, and confirm password fields are not empty and that the password and confirm password fields match. The validation logic uses `isEmpty()` and sets error messages on the corresponding input fields if there is an error.

* Use view binding to bind views in SignupFragment and handle signup logic.

* Add functionality to get user by email

Added a new query in `UserDao` to retrieve a user based on their email.
Created a new `UserRepository` class with a `getUserByEmail` method that uses the new query in `UserDao`.

* feat: Implement login functionality

This commit introduces the login feature. It includes the creation of a `LoginUseCase` to handle the login logic, a `LoginViewModel` to manage the UI state and interact with the use case, and the necessary dependency injection configurations in `AppModule`. It also introduces the new `UserRepository` to fetch user from the local database. The `LandingViewModel` is now `LoginViewModel` and handles the login functionality.

* Refactor: Use Hilt for LoginViewModel and LoginFragment

- Updated `LoginViewModel` to use `@HiltViewModel` and constructor injection.
- Injected `LoginUseCase` into `LoginViewModel`.
- Implemented actual login logic with `LoginUseCase` in `LoginViewModel`.
- Added `@AndroidEntryPoint` to `LoginFragment` to enable Hilt injection.
- Removed unnecessary `LoginViewModel` code from `LandingViewModel` class.

* Refactor: Implement Login Use Case and ViewModel

- Inject `LoginUseCase` into `LoginViewModel`.
- Call `LoginUseCase` from `LoginViewModel` with email and password.
- Implement `LoginUseCase` to call `UserRepository`'s `getUserByEmail`.
- Return true in `LoginUseCase` if user exists, otherwise false.