Feature/database logic/nathan (#30)
* Add a foreign key for category and update TransactionEntity

The TransactionEntity class was updated to add a foreign key constraint referencing the CategoryEntity.
Also TransactionEntity class has been updated by adding new columns.
The selected tab id for the app insights was set to Firebase Crashlytics.

* Define CategoryEntity as a Room entity with foreign key to UserEntity.

* feat: Add BudgetEntity and BudgetDao

- Add `BudgetEntity` with foreign keys to `UserEntity` and `WalletEntity`.
- Add `BudgetDao` with methods to insert a budget, get a budget by ID, get budgets by user ID, and get budgets by wallet ID.
- Update `CategoryEntity` to use the `ForeignKey` annotation.

* Update AppDatabase entities and version

Add BudgetEntity to the list of entities for the database.
Increment the database version from 2 to 3.