feat: Add CategoryDatabaseCallback and default categories
This commit introduces the `CategoryDatabaseCallback` class, which automatically inserts default categories into the database when it's first created. The `CategoryDefualts` object defines the list of these default categories.
The `CategoryEntity` has been updated to allow null `user_id`, to enable default categories to be globally accessible. `CategoryDao` now uses `insertCategory` method.
The `AppModule` was modified to add the callback to the room database builder.
The database version was incremented to `4`.