Feature/label table/nathan (#32)
* Add LabelDao and LabelEntity

- Introduce LabelEntity for label data management.
- Create LabelDao with methods for inserting, retrieving by ID, and retrieving all user-specific labels.
- Add foreign key constraint from CategoryEntity to UserEntity.

* feat: Add CategoryDatabaseCallback and default categories

This commit introduces the `CategoryDatabaseCallback` class, which automatically inserts default categories into the database when it's first created. The `CategoryDefualts` object defines the list of these default categories.
The `CategoryEntity` has been updated to allow null `user_id`, to enable default categories to be globally accessible. `CategoryDao` now uses `insertCategory` method.
The `AppModule` was modified to add the callback to the room database builder.
The database version was incremented to `4`.

* Refactor: Move and rename CategoryDefualts, add fallbackToDestructiveMigration

- Moved `CategoryDefualts` from `data.defuaults` to `data._default`.
- Added `fallbackToDestructiveMigration()` to the database builder to allow for easier migration during testing and development.
- Added comment for `CategoryDefualts` to explain default categories.