feat: Add Budget model, entity, and mappers
This commit introduces the `Budget` model and its corresponding `BudgetEntity` for database representation. It also adds mappers to convert between `BudgetIn`, `BudgetEntity`, and `Budget` domain objects. Additionally, it modifies the `BudgetIn` and `Budget` classes, removing optionality from budgetId, and adds the `BudgetWithUser` relation to manage the link between `Budget` and `User` in database.