Feature/adding new icons/nathan (#91)
* feat: Add new vector drawables for UI elements and categories

This commit introduces a collection of new vector drawable assets.

New icons include:
- `ic_cat_art.xml`
- `ic_cat_education.xml`
- `ic_cat_electronics.xml`
- `ic_cat_food.xml`
- `ic_cat_fuel.xml`
- `ic_cat_pet.xml`
- `ic_cat_travel.xml`
- `ic_categories.xml`
- `ic_linegraph.xml`
- `ic_piechart.xml`
- `ic_privacy_close.xml`
- `ic_ui_circle_check.xml`
- `ic_ui_notebook.xml`

These icons will be used for various UI elements and category representations within the application. Each drawable is defined with paths and attributes for rendering, and they are styled to use the `bb_primaryText` color attribute.

* feat: Add new vector drawables for UI elements and categories

This commit introduces a collection of new vector drawable assets.

New icons include:
- `ic_cat_art.xml`
- `ic_cat_education.xml`
- `ic_cat_electronics.xml`
- `ic_cat_food.xml`
- `ic_cat_fuel.xml`
- `ic_cat_pet.xml`
- `ic_cat_travel.xml`
- `ic_categories.xml`
- `ic_linegraph.xml`
- `ic_piechart.xml`
- `ic_privacy_close.xml`
- `ic_ui_circle_check.xml`
- `ic_ui_notebook.xml`

These icons will be used for various UI elements and category representations within the application. Each drawable is defined with paths and attributes for rendering, and they are styled to use the `bb_primaryText` color attribute.